TODO (1) implement LoaderManager.LoaderCallbacks<String> on MainActivity TODO (2) Create a constant int to uniquely identify your loader. Call it GITHUB_SEARCH_LOADER TODO (3) Override onCreateLoader // Within onCreateLoader     // TODO (4) Return a new AsyncTaskLoader<String> as an anonymous inner class with this as the constructor's parameter         // TODO (5) Override onStartLoading             // Within onStartLoading             // TODO (6) If args is null, return.             // TODO (7) Show the loading indicator             // TODO (8) Force a load             // END - onStartLoading
        // TODO (9) Override loadInBackground
            // Within loadInBackground
            // TODO (10) Get the String for our URL from the bundle passed to onCreateLoader
            // TODO (11) If the URL is null or empty, return null
            // TODO (12) Copy the try / catch block from the AsyncTask's doInBackground method
            // END - loadInBackground

// TODO (13) Override onLoadFinished

    // Within onLoadFinished
    // TODO (14) Hide the loading indicator
    // TODO (15) Use the same logic used in onPostExecute to show the data or the error message
    // END - onLoadFinished

TODO (16) Override onLoaderReset as it is part of the interface we implement, but don't do anything in this method
TODO (17) If no search was entered, indicate that there isn't anything to search for and return
TODO (18) Remove the call to execute the AsyncTask
TODO (19) Create a bundle called queryBundle
TODO (20) Use putString with SEARCH_QUERY_URL_EXTRA as the key and the String value of the URL as the value
TODO (21) Call getSupportLoaderManager and store it in a LoaderManager variable
TODO (22) Get our Loader by calling getLoader and passing the ID we specified
TODO (23) If the Loader was null, initialize it. Else, restart it.
TODO (24) Initialize the loader with GITHUB_SEARCH_LOADER as the ID, null for the bundle, and this for the context
TODO (25) Remove the code that displays the JSON
TODO (26) Remove the code that retrieves the JSON
TODO (27) Remove the code that persists the JSON
TODO (28) Remove the key for storing the search results JSON
TODO (29) Delete the AsyncTask class